Fixed test import issues

+++ serverextension/test/test_storage.py
@@ -5,19 +5,20 @@ test the DbHandler methods
 import unittest
 import sqlite3
 import os
+from context import prompter
 
 class TestDBMethods(unittest.TestCase):
 
     def setUp(self):
 
-        self.TEST_DB_DIR = "./test/"
+        self.TEST_DB_DIR = "./"
         self.TEST_DB_NAME = "cellstest.db"
 
     def test_init(self):
 
-        db = prompter.storage.DbHandler(dirname=TEST_DB_DIR, dbname=TEST_DB_NAME)
+        db = prompter.DbHandler(dirname=self.TEST_DB_DIR, dbname=self.TEST_DB_NAME)
         
-        conn = sqlite3.connect(TEST_DB_DIR + TEST_DB_NAME)
+        conn = sqlite3.connect(self.TEST_DB_DIR + self.TEST_DB_NAME)
         cursor = conn.cursor()
 
         cursor.execute(
@@ -39,9 +40,8 @@ class TestDBMethods(unittest.TestCase):
         self.assertTrue("versions" in table_names)
 
     def tearDown(self):
-
-        os.remove(self.TEST_DB_DIR+self.TEST_DB_NAME)
-        os.rmdir(self.TEST_DB_DIR)
+        if os.is_file(self.TEST_DB_DIR+self.TEST_DB_NAME):
+            os.remove(self.TEST_DB_DIR+self.TEST_DB_NAME)
 
 if __name__ == "__main__":
     unittest.main()
